+++ CHANGELOG.md
@@ -17,6 +17,7 @@ changes in the following format: PR #1234***
 Quality Control, and Behavioural Quality Control. (PR #6041)
 - Addition of a new `account_request_date` in `users` table that will be used when
 requesting a new account and will be displayed in the User Accounts module (PR #6191)
+- Candidate's age can be retrieved from the Candidate class in days, months, or years (PR #5945)
 #### Bug Fixes
 - *Add item here*
 ### Modules